DamCTF2023/thunderstruck/chall/requirements.txt
quart==0.18.3
quart-auth==0.8.0
hypercorn==0.14.3
aiofiles==23.1.0
synapse==2.126.0
pycryptodome==3.16.0